Power bi Project
Road Accidents Analysis reveals the series of accidents as they occurred according to Road Type, by Locations and by Vehicle Type. It also shows the monthly trend of these accidents and it's impacts, if it is fatal, serious or slight. The dashboards equally shows the effect of accidents by comparing it based on rurual/urban and by Light Conditions(Night and Day)